Stop reading 0001-01-01 as NULL Timestamp
Over two years ago, pantab had a bug due to which `pandas.NaT` was
written to Hyper files as "0001-01-01".
Consequently, the reader was also written to interpret "0001-01-01"
as NULL. This was never quite correct, but was kept as is to not break
backwards compatibility.

Now, two years later, it's time to remove that hack.

closes #64